feat(sdk-metrics)!: drop deprecated `InstrumentDescriptor` export
Note that this public `InstrumentDescriptor` has always been just
an alias of `MetricDescriptor`, and does not actually point to the
internal type with the same name.

Separately:

* Refactor the internal `InstrumentDescriptor` type to extend from
  the public `MetricDescriptor`, adding only the `advice` field

* Move the `InstrumentType` enum into `./src/export/MetricData.ts`
  as it is a public export, plus to avoid a circular dependency
  after the refactor
