+++ packages/sdk-metrics/src/Meter.ts
@@ -27,10 +27,7 @@ import {
   BatchObservableCallback,
   Observable,
 } from '@opentelemetry/api';
-import {
-  createInstrumentDescriptor,
-  InstrumentType,
-} from './InstrumentDescriptor';
+import { createInstrumentDescriptor } from './InstrumentDescriptor';
 import {
   CounterInstrument,
   GaugeInstrument,
@@ -41,6 +38,7 @@ import {
   UpDownCounterInstrument,
 } from './Instruments';
 import { MeterSharedState } from './state/MeterSharedState';
+import { InstrumentType } from './export/MetricData';
 
 /**
  * This class implements the {@link IMeter} interface.
