+++ packages/sdk-metrics/src/index.ts
@@ -14,8 +14,6 @@
  * limitations under the License.
  */
 
-import { MetricDescriptor } from './export/MetricData';
-
 export {
   Sum,
   LastValue,
@@ -36,6 +34,7 @@ export {
   SumMetricData,
   GaugeMetricData,
   HistogramMetricData,
+  InstrumentType,
   ExponentialHistogramMetricData,
   ResourceMetrics,
   ScopeMetrics,
@@ -58,12 +57,6 @@ export { InMemoryMetricExporter } from './export/InMemoryMetricExporter';
 export { ConsoleMetricExporter } from './export/ConsoleMetricExporter';
 
 export { MetricCollectOptions, MetricProducer } from './export/MetricProducer';
-
-export { InstrumentType } from './InstrumentDescriptor';
-/**
- * @deprecated Use {@link MetricDescriptor} instead.
- */
-export type InstrumentDescriptor = MetricDescriptor;
 
 export { MeterProvider, MeterProviderOptions } from './MeterProvider';
